Move time.hpp out of include/deepstream
